# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tasks API - OpenAPI 3.0",
    "description": "This a simple API to administer your tasks.\n\nSome useful links:\n- [The API repository](https://github.com/ViniciusBessa/football-api)\n- [The source API definition used](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "contact": {
      "email": "vinicius.b.silva3@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": ["task"],
        "summary": "Get all tasks",
        "description": "Get all tasks",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "Successfully retrieved all tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["task"],
        "summary": "Create a task",
        "description": "Create a task",
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "description": "Create a task in the database",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Brazil"
                  },
                  "code": {
                    "type": "string",
                    "example": "BR"
                  },
                  "flagUrl": {
                    "type": "string",
                    "example": "https://flags.com/brazil.jpg"
                  }
                },
                "required": ["name", "code", "flagUrl"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to create a task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "It's necessary to login to see this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "You don't have the permission to access this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/tasks/{taskId}": {
      "get": {
        "tags": ["task"],
        "summary": "Get a specific task",
        "description": "Get a specific task",
        "operationId": "getTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "ID of task to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No task was found with the provided id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["task"],
        "summary": "Update a task",
        "description": "Update a task",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "ID of task to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Update a specific task",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Brazil"
                  },
                  "code": {
                    "type": "string",
                    "example": "BR"
                  },
                  "flagUrl": {
                    "type": "string",
                    "example": "https://flags.com/brazil.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    },
                    "token": {
                      "$ref": "#/components/schemas/Token"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to update a task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "It's necessary to login to see this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "You don't have the permission to access this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "No task was found with the provided id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["task"],
        "summary": "Delete a task",
        "description": "Delete a task",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "ID of task to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "It's necessary to login to see this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "You don't have the permission to access this content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "404": {
            "description": "No task was found with the provided id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Robert"
          },
          "email": {
            "type": "string",
            "example": "robert@email.com"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": ["USER", "ADMIN"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "1"
          },
          "title": {
            "type": "string",
            "example": "My First Task"
          },
          "description": {
            "type": "string",
            "example": "The first task of the day"
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "user": {
            "type": "User",
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Token": {
        "type": "string",
        "example": "jwtToken"
      },
      "BadRequestError": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "example": "Something is wrong with the request"
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "example": "It's necessary to login to see this content"
          }
        }
      },
      "ForbiddenError": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "example": "You don't have the permission to access this content"
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "example": "No content was found with the provided id"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}